Exercise7.jl Documentation

Compartmental Models

Exercise7.AbstractCompartmentalModel — Type

AbstractCompartmentalModel{T<:AbstractFloat}

Supertype of all compartmental models defined in this package.

SIR Models

Exercise7.AbstractSIRModel — Type

AbstractSIRModel{T} <: AbstractCompartmentalModel{T}

Supertype of all SIR models.

Exercise7.SIR — Type

SIR{T} <: AbstractSIRModel{T}

Standard SIR model with contact rate $\beta$, recovery rate $\gamma$, and no demography:

\[\begin{aligned} \frac{dS}{dt} & = -\beta I S \\[8pt] \frac{dI}{dt} & = \beta I S - \gamma I \\[8pt] \frac{dR}{dt} & = \gamma I . \end{aligned}\]

Objects of this type are callable with the signature (du, u, p, t) which performs an in-place update of the derivatives du of the system.

Exercise7.SIR — Method

(::SIR)(du, u, p, t) -> nothing

Update the derivates du given a model of type SIR.

Exercise7.SIR — Method

SIR() -> SIR

Return an SIR model with default parameters $\beta = 0.35$ and $\gamma = 0.035$.

Exercise7.SIRDecayingImmunity — Type

SIRDecayingImmunity{T} <: AbstractSIRModel{T}

SIR model with contact rate $\beta$, recovery rate $\gamma$, infection immunity decay rate $\alpha$, and no demography:

\[\begin{aligned} \frac{dS}{dt} & = -\beta I S + \alpha R \\[8pt] \frac{dI}{dt} & = \beta I S - \gamma I \\[8pt] \frac{dR}{dt} & = \gamma I - \alpha R . \end{aligned}\]

Objects of this type are callable with the signature (du, u, p, t) which performs an in-place update of the derivatives du of the system.

Exercise7.SIRDecayingImmunity — Method

SIRDecayingImmunity(α) -> SIRDecayingImmunity

Return an SIRDecayingImmunity model with immunity decay rate $\alpha$ and default parameters $\beta = 0.35$ and $\gamma = 0.035$.

Exercise7.SIRDecayingImmunity — Method

(::SIRDecayingImmunity)(du, u, p, t) -> nothing

Update the derivates du given a model of type SIRDecayingImmunity.

SIRV Models

Exercise7.AbstractSIRVModel — Type

AbstractSIRVModel{T} <: AbstractCompartmentalModel{T}

Supertype of all SIRV models.

Exercise7.SIRV — Type

SIRV{T} <: AbstractSIRVModel{T}

SIRV model with contact rate $\beta$, recovery rate $\gamma$, vaccination rate $\nu$, and no demography:

\[\begin{aligned} \frac{dS}{dt} & = -\beta I S -\nu S\\[8pt] \frac{dI}{dt} & = \beta I S - \gamma I \\[8pt] \frac{dR}{dt} & = \gamma I \\[8pt] \frac{dV}{dt} & = \nu S . \end{aligned}\]

Objects of this type are callable with the signature (du, u, p, t) which performs an in-place update of the derivatives du of the system.

Exercise7.SIRV — Method

(::SIRV)(du, u, p, t) -> nothing

Update the derivates du given a model of type SIRV.

Exercise7.SIRV — Method

SIRV(ν) -> SIRV

Return an SIRV model with vaccination rate $\nu$ and default parameters $\beta = 0.35$ and $\gamma = 0.035$.

Exercise7.SIRVSeasonalContact — Type

SIRVSeasonalContact{T} <: AbstractSIRVModel{T}

SIRV model with seasonal contact rate $\beta(t)$, recovery rate $\gamma$, vaccination rate $\nu$, and no demography.

The contact rate $\beta(t)$ is given by $\beta(t) = \beta_0(1 + \beta_1 cos(2\pi t / 365))$, where $\beta_0$ is the mean contact rate and $0 \leq \beta_1 \leq 1$ controls the degree of seasonality.

The full equations of motion are:

\[\begin{aligned} \frac{dS}{dt} & = -\beta(t) I S -\nu S\\[8pt] \frac{dI}{dt} & = \beta(t) I S - \gamma I \\[8pt] \frac{dR}{dt} & = \gamma I \\[8pt] \frac{dV}{dt} & = \nu S . \end{aligned}\]

Objects of this type are callable with the signature (du, u, p, t) which performs an in-place update of the derivatives du of the system.

Exercise7.SIRVSeasonalContact — Method

SIRVSeasonalContact(β₁, ν) -> SIRVSeasonalContact

Return an SIRVSeasonalContact model with seasonality parameter $\beta_1$, vaccination rate $\nu$, and default parameters $\beta_0 = 0.35$ and $\gamma = 0.035$.

Exercise7.SIRVSeasonalContact — Method

(::SIRVSeasonalContact)(du, u, p, t) -> nothing

Update the derivates du given a model of type SIRVSeasonalContact.

Exercise7.SIRVDecayingImmunity — Type

SIRVDecayingImmunity{T} <: SIRVDecayingImmunity{T}

SIRV model with contact rate $\beta$, recovery rate $\gamma$, vaccination rate $\nu$, infection immunity decay rate $\alpha$, vaccination immunity decay rate $\mu$, and no demography:

\[\begin{aligned} \frac{dS}{dt} & = -\beta I S -\nu S + \alpha R + \mu V\\[8pt] \frac{dI}{dt} & = \beta I S - \gamma I \\[8pt] \frac{dR}{dt} & = \gamma I - \alpha R\\[8pt] \frac{dV}{dt} & = \nu S - \mu V. \end{aligned}\]

Objects of this type are callable with the signature (du, u, p, t) which performs an in-place update of the derivatives du of the system.

Exercise7.SIRVDecayingImmunity — Method

SIRVDecayingImmunity(ν, α, μ) -> SIRVDecayingImmunity

Return an SIRVDecayingImmunity model with vaccination rate $\nu$, infection immunity decay rate $\alpha$, vaccination immunity decay rate $\mu$, and default parameters $\beta_0 = 0.35$ and $\gamma = 0.035$.

Exercise7.SIRVDecayingImmunity — Method

(::SIRVDecayingImmunity)(du, u, p, t) -> nothing

Update the derivates du given a model of type SIRVDecayingImmunity.

Exercise7.SIRVSeasonalContactDecayingImmunity — Type

SIRVSeasonalContactDecayingImmunity{T} <: SIRVSeasonalContactDecayingImmunity{T}

SIRV model with seasonal contact rate $\beta(t)$, recovery rate $\gamma$, vaccination rate $\nu$, infection immunity decay rate $\alpha$, vaccination immunity decay rate $\mu$, and no demography.

The contact rate $\beta(t)$ is given by $\beta(t) = \beta_0(1 + \beta_1 cos(2\pi t / 365))$, where $\beta_0$ is the mean contact rate and $0 \leq \beta_1 \leq 1$ controls the degree of seasonality.

The full equations of motion are:

\[\begin{aligned} \frac{dS}{dt} & = -\beta(t) I S -\nu S + \alpha R + \mu V\\[8pt] \frac{dI}{dt} & = \beta(t) I S - \gamma I \\[8pt] \frac{dR}{dt} & = \gamma I - \alpha R\\[8pt] \frac{dV}{dt} & = \nu S - \alpha V. \end{aligned}\]

Objects of this type are callable with the signature (du, u, p, t) which performs an in-place update of the derivatives du of the system.

Exercise7.SIRVSeasonalContactDecayingImmunity — Method

(::SIRVSeasonalContactDecayingImmunity)(du, u, p, t) -> nothing

Update the derivates du given a model of type SIRVSeasonalContactDecayingImmunity.